+++ src/test/java/seedu/address/commons/events/ui/SwitchToEventsListEventTest.java
@@ -1,9 +1,10 @@
-package seedu.address.commons.events;
+package seedu.address.commons.events.ui;
 
 import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
 
+import seedu.address.commons.events.BaseEvent;
 import seedu.address.commons.events.ui.SwitchToEventsListEvent;
 
 public class SwitchToEventsListEventTest {
